requirements.txt
django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
environs[django]==14.1.1